Se agrega operacion cambiarDeNivelOp, se corrige problemas al comer elementos. Ahora puede cambiar de mapa hay ver preparar los clientes para que reciban bien los mensajes

+++ server/src/Config.cpp
@@ -30,7 +30,7 @@ void Config::setInstance(string mundoXmlPath,int port,int minJugadores,
     }
 }
 
-Config::Config(string mundoXmlPath,int port,int minJugadores,int maxJugadores,
+Config::Config(string mundoXmlPath,int port,size_t minJugadores,size_t maxJugadores,
 	           bool com_fantasmas,int vidas,string archivoLog){
 	this->mundoXmlPath=mundoXmlPath;
 	this->port=port;
